Moved ORProjectManager::findAncestorRecursive --> ORFileUtils::findAncestorRecursive.
