+++ src/com/reason/ide/ORProjectManager.java
@@ -34,33 +34,6 @@ public class ORProjectManager {
 
     private ORProjectManager() {}
 
-    public static Optional<VirtualFile> findAncestorRecursive(@NotNull Project project,
-            @NotNull String target, @NotNull VirtualFile start) {
-        // start must be a directory, should only happen on first iteration
-        if (!start.isDirectory()) {
-           start = start.getParent();
-            // parent is null, done
-           if (start == null) {
-               return Optional.empty();
-           }
-        }
-        // target found, done
-        VirtualFile foundTarget = start.findChild(target);
-        if (foundTarget != null) {
-            return Optional.of(foundTarget);
-        }
-        // just checked project root, done
-        if (start.getPath().equals(project.getBasePath())) {
-            return Optional.empty();
-        }
-        // parent is null, we're done
-        VirtualFile parent = start.getParent();
-        if (parent == null) {
-            return Optional.empty();
-        }
-        return findAncestorRecursive(project, target, parent);
-    }
-
     public static boolean isBsProject(@NotNull Project project) {
         return !findBsConfigurationFiles(project).isEmpty();
     }
